+++ src/asp/Tools/stereo_pprc.cc
@@ -296,8 +296,7 @@ void stereo_preprocessing(bool adjust_left_image_size, ASPGlobalOptions& opt) {
            has_left_georef, left_georef,
            false, output_nodata,
            opt, TerminalProgressCallback
-           ("asp", "\t  StdDev filter raw output (left): ")
-           );
+           ("asp", "\t  StdDev filter raw output (left): "));
         vw::cartography::block_write_gdal_image
           (opt.out_prefix + "-R_stddev_filter_output.tif",
            vw::stddev_filter_view(right_image, stddev_kernel),