+++ NEWS.rst
@@ -1,3 +1,12 @@
+Changes since last release
+--------------------------
+
+stereo:
+
+  * Document pre-processing options ``--stddev-mask-kernel``
+    and ``--stddev-mask-thresh`` (:numref:`stereo-default-preprocessing`).
+    Also fix a bug in writing out helper debug images for this option.
+ 
 RELEASE 3.2.0, December 30, 2022
 --------------------------------
 DOI: `10.5281/zenodo.7497499 <https://doi.org/10.5281/zenodo.7497499>`_